+++ src/pools/utils.ts
@@ -18,6 +18,18 @@ import {
 import type { MessageValue, Task } from '../utility-types.ts'
 import type { WorkerChoiceStrategyContext } from './selection-strategies/worker-choice-strategy-context.ts'
 
+export const getDefaultTasksQueueOptions = (
+  poolMaxSize: number,
+): Required<TasksQueueOptions> => {
+  return {
+    size: Math.pow(poolMaxSize, 2),
+    concurrency: 1,
+    taskStealing: true,
+    tasksStealingOnBackPressure: true,
+    tasksFinishedTimeout: 1000,
+  }
+}
+
 export const checkFileURL = (fileURL: URL): void => {
   if (fileURL == null) {
     throw new TypeError('The worker URL must be specified')
@@ -320,6 +332,7 @@ export const waitWorkerNodeEvents = async <
   workerNode: IWorkerNode<Worker, Data>,
   workerNodeEvent: string,
   numberOfEventsToWait: number,
+  timeout: number,
 ): Promise<number> => {
   return await new Promise<number>((resolve) => {
     let events = 0
@@ -333,5 +346,10 @@ export const waitWorkerNodeEvents = async <
         resolve(events)
       }
     })
+    if (timeout > 0) {
+      setTimeout(() => {
+        resolve(events)
+      }, timeout)
+    }
   })
 }