+++ tests/pools/abstract-pool.test.mjs
@@ -314,6 +314,7 @@ Deno.test({
           size: Math.pow(numberOfWorkers, 2),
           taskStealing: true,
           tasksStealingOnBackPressure: true,
+          tasksFinishedTimeout: 1000,
         },
         workerChoiceStrategy: WorkerChoiceStrategies.LEAST_USED,
         workerChoiceStrategyOptions: {
@@ -769,6 +770,7 @@ Deno.test({
           size: Math.pow(numberOfWorkers, 2),
           taskStealing: true,
           tasksStealingOnBackPressure: true,
+          tasksFinishedTimeout: 1000,
         })
         pool.enableTasksQueue(true, { concurrency: 2 })
         expect(pool.opts.enableTasksQueue).toBe(true)
@@ -777,6 +779,7 @@ Deno.test({
           size: Math.pow(numberOfWorkers, 2),
           taskStealing: true,
           tasksStealingOnBackPressure: true,
+          tasksFinishedTimeout: 1000,
         })
         pool.enableTasksQueue(false)
         expect(pool.opts.enableTasksQueue).toBe(false)
@@ -801,6 +804,7 @@ Deno.test({
           size: Math.pow(numberOfWorkers, 2),
           taskStealing: true,
           tasksStealingOnBackPressure: true,
+          tasksFinishedTimeout: 1000,
         })
         for (const workerNode of pool.workerNodes) {
           expect(workerNode.tasksQueueBackPressureSize).toBe(
@@ -812,12 +816,14 @@ Deno.test({
           size: 2,
           taskStealing: false,
           tasksStealingOnBackPressure: false,
+          tasksFinishedTimeout: 2000,
         })
         expect(pool.opts.tasksQueueOptions).toStrictEqual({
           concurrency: 2,
           size: 2,
           taskStealing: false,
           tasksStealingOnBackPressure: false,
+          tasksFinishedTimeout: 2000,
         })
         for (const workerNode of pool.workerNodes) {
           expect(workerNode.tasksQueueBackPressureSize).toBe(
@@ -834,6 +840,7 @@ Deno.test({
           size: Math.pow(numberOfWorkers, 2),
           taskStealing: true,
           tasksStealingOnBackPressure: true,
+          tasksFinishedTimeout: 1000,
         })
         for (const workerNode of pool.workerNodes) {
           expect(workerNode.tasksQueueBackPressureSize).toBe(